renaming and simplification
