+++ processing.py
@@ -4,8 +4,6 @@ __author__ = 'Holger Stitz'
 from flask import Flask, Response
 from caleydo_server.util import jsonify
 
-from caleydo_processing_queue.celery import notifier, get_result
-
 import logging
 _log = logging.getLogger(__name__)
 
@@ -17,7 +15,7 @@ def stream():
   event source like stream for requesting task results
   :return:
   """
-  from caleydo_processing_queue.celery import notifier
+  from caleydo_processing_queue.queue import notifier
   import json
   def event_stream():
     for msg in notifier.listen():
@@ -33,6 +31,7 @@ def get_result(task_id):
   :param task_id:
   :return:
   """
+  from caleydo_processing_queue.queue import get_result
   res = get_result(task_id)
   return jsonify(res.get())
